Move components to separate folder
