+++ typescript/components/WebSocketNext.tsx
@@ -1,5 +1,5 @@
 import React, { FC } from 'react'
-import { WebSocketProps } from './types/WebSocket.props'
+import { WebSocketProps } from '../types/WebSocket.props'
 
 const WebSocketNext: FC<WebSocketProps> = (props) => {
   return <div {...props} />
